[Consolidation] Migrate Exception Classes to Own
