+++ halpybot/commands/caseutils.py
@@ -1,5 +1,5 @@
 """
-caseutils.py - Commands involving the management of Seal cases
+CommandUtils.py - Commands involving the management of Seal cases
 
 Copyright (c) The Hull Seals,
 All rights reserved.
@@ -7,22 +7,22 @@ All rights reserved.
 Licensed under the GNU General Public License
 See license.md
 """
-import functools
 import re
 from typing import List
 from pendulum import now
 from halpybot import config
-from ..packages.command import Commands, get_help_text
-from ..packages.edsm import (
+from ..packages.exceptions import NoUserFound, NoResultsEDSM, EDSMConnectionError
+from ..packages.utils import (
+    CommandUtils,
     sys_cleaner,
+)
+from ..packages.command import Commands
+from ..packages.edsm import (
     checklandmarks,
-    NoResultsEDSM,
-    EDSMConnectionError,
 )
 from ..packages.models import (
     Context,
     User,
-    NoUserFound,
     Case,
     Status,
     CaseType,
@@ -31,33 +31,7 @@ from ..packages.models import (
     KFType,
 )
 from ..packages.checks import Require, Drilled
-from ..packages.case import get_case, update_single_elem_case_prep
-
-
-# Decorators
-class CaseUtils:
-    """Utilities for Wrapping EDSM Commands"""
-
-    @staticmethod
-    def gather_case(len_args_expected: int):
-        """Process gathering Case details for a command expecting a number of args"""
-
-        def decorator(function):
-            @functools.wraps(function)
-            async def guarded(ctx, args: List[str]):
-                if len(args) < len_args_expected:
-                    return await ctx.reply(
-                        get_help_text(ctx.bot.commandsfile, ctx.command)
-                    )
-                try:
-                    case: Case = await get_case(ctx, args[0])
-                except KeyError:
-                    return await ctx.reply(f"No case found for {args[0]!r}.")
-                return await function(ctx, args, case)
-
-            return guarded
-
-        return decorator
+from ..packages.case import update_single_elem_case_prep
 
 
 # FACT WRAPPERS
@@ -162,7 +136,7 @@ async def cmd_listboard(ctx: Context, args: List[str]):
 
 @Commands.command("listcase")
 @Require.permission(Drilled)
-@CaseUtils.gather_case(1)
+@CommandUtils.gather_case(1)
 async def cmd_listcase(ctx: Context, args: List[str], case: Case):
     """
     Send a user the key details of a case on the board in DMs
@@ -177,7 +151,7 @@ async def cmd_listcase(ctx: Context, args: List[str], case: Case):
 @Commands.command("rename")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_renamecase(ctx: Context, args: List[str], case: Case):
     """
     Rename the user of an active case
@@ -199,7 +173,7 @@ async def cmd_renamecase(ctx: Context, args: List[str], case: Case):
 @Commands.command("ircn")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_ircn(ctx: Context, args: List[str], case: Case):
     """
     Rename the user of an active case
@@ -221,7 +195,7 @@ async def cmd_ircn(ctx: Context, args: List[str], case: Case):
 @Commands.command("system")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_system(ctx: Context, args: List[str], case: Case):
     """
     Change the system of an active case
@@ -250,7 +224,7 @@ async def cmd_system(ctx: Context, args: List[str], case: Case):
 @Commands.command("status")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_status(ctx: Context, args: List[str], case: Case):
     """
     Change the activity status of a case
@@ -279,7 +253,7 @@ async def cmd_status(ctx: Context, args: List[str], case: Case):
 @Commands.command("hull")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_hull(ctx: Context, args: List[str], case: Case):
     """
     Change the starting hull percentage of a case
@@ -309,7 +283,7 @@ async def cmd_hull(ctx: Context, args: List[str], case: Case):
 @Commands.command("changetype")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_changetype(ctx: Context, args: List[str], case: Case):
     """
     Change the case type between Seal, KF, Black, or Blue.
@@ -348,7 +322,7 @@ async def cmd_changetype(ctx: Context, args: List[str], case: Case):
 @Commands.command("platform")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_platform(ctx: Context, args: List[str], case: Case):
     """
     Change the platform a case is on.
@@ -391,7 +365,7 @@ async def cmd_platform(ctx: Context, args: List[str], case: Case):
 @Commands.command("planet")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_planet(ctx: Context, args: List[str], case: Case):
     """
     Change the planet of an active KF case
@@ -412,7 +386,7 @@ async def cmd_planet(ctx: Context, args: List[str], case: Case):
 @Commands.command("casecoords")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(3)
+@CommandUtils.gather_case(3)
 async def cmd_coords(ctx: Context, args: List[str], case: Case):
     """
     Change the coords of an active KF case
@@ -442,7 +416,7 @@ async def cmd_coords(ctx: Context, args: List[str], case: Case):
 @Commands.command("o2time")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_oxtime(ctx: Context, args: List[str], case: Case):
     """
     Change the remaining oxygen timer for a case
@@ -469,7 +443,7 @@ async def cmd_oxtime(ctx: Context, args: List[str], case: Case):
 @Commands.command("synth")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_synth(ctx: Context, args: List[str], case: Case):
     """
     Toggle if a CMDR has synths available for a given case.
@@ -499,7 +473,7 @@ async def cmd_synth(ctx: Context, args: List[str], case: Case):
 @Commands.command("canopy")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_canopy(ctx: Context, args: List[str], case: Case):
     """
     Toggle if the canopy is broken for a given case.
@@ -530,7 +504,7 @@ async def cmd_canopy(ctx: Context, args: List[str], case: Case):
 @Commands.command("kftype")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_changekftype(ctx: Context, args: List[str], case: Case):
     """
     Change the case type between KF subtypes.
@@ -566,7 +540,7 @@ async def cmd_changekftype(ctx: Context, args: List[str], case: Case):
 @Commands.command("notes", "updatenotes", "addnote")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_notes(ctx: Context, args: List[str], case: Case):
     """
     Append a new entry to the Notes
@@ -583,7 +557,7 @@ async def cmd_notes(ctx: Context, args: List[str], case: Case):
 
 @Commands.command("listnotes")
 @Require.permission(Drilled)
-@CaseUtils.gather_case(1)
+@CommandUtils.gather_case(1)
 async def cmd_listnotes(ctx: Context, args: List[str], case: Case):
     """
     List out just the case notes to DMs
@@ -602,7 +576,7 @@ async def cmd_listnotes(ctx: Context, args: List[str], case: Case):
 @Commands.command("delnote")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(2)
+@CommandUtils.gather_case(2)
 async def cmd_delnote(ctx: Context, args: List[str], case: Case):
     """
     Delete a line of the notes for a given case
@@ -628,7 +602,7 @@ async def cmd_delnote(ctx: Context, args: List[str], case: Case):
 @Commands.command("editnote")
 @Require.permission(Drilled)
 @Require.channel()
-@CaseUtils.gather_case(3)
+@CommandUtils.gather_case(3)
 async def cmd_editnote(ctx: Context, args: List[str], case: Case):
     """
     Alter a line of the notes for a given case
