+++ halpybot/packages/announcer/__init__.py
@@ -11,23 +11,15 @@ See license.md
 from .announcer import (
     Announcer,
     Announcement,
-    AnnouncementError,
     get_edsm_data,
-    AlreadyExistsError,
     AnnouncerArgs,
-    KFCoordsError,
 )
-from .dc_webhook import send_webhook, DiscordWebhookError, WebhookSendError
+from .dc_webhook import send_webhook
 
 __all__ = [
     "Announcer",
     "Announcement",
-    "AnnouncementError",
     "get_edsm_data",
-    "AlreadyExistsError",
     "AnnouncerArgs",
-    "KFCoordsError",
     "send_webhook",
-    "DiscordWebhookError",
-    "WebhookSendError",
 ]
