+++ halpybot/packages/edsm/__init__.py
@@ -11,35 +11,23 @@ See license.md
 from .edsm import (
     GalaxySystem,
     Commander,
-    EDSMLookupError,
-    EDSMConnectionError,
     checkdistance,
     checkdssa,
     checklandmarks,
     get_nearby_system,
-    sys_cleaner,
-    NoResultsEDSM,
     calc_distance,
     calc_direction,
     diversions,
-    NoNearbyEDSM,
-    EDSMReturnError,
 )
 
 __all__ = [
     "GalaxySystem",
     "Commander",
-    "EDSMLookupError",
-    "EDSMConnectionError",
     "checklandmarks",
     "checkdssa",
     "checkdistance",
     "get_nearby_system",
-    "sys_cleaner",
-    "NoResultsEDSM",
     "calc_distance",
     "calc_direction",
     "diversions",
-    "NoNearbyEDSM",
-    "EDSMReturnError",
 ]
