+++ halpybot/packages/models/__init__.py
@@ -9,7 +9,7 @@ See license.md
 """
 
 from .edsm_classes import Coordinates, Location, SystemInfo
-from .user import User, UserError, NoUserFound
+from .user import User
 from .context import Context, HelpArguments
 from .case import Case, Platform, KFCoords, Status, CaseType, KFType
 from .seal import Seal
@@ -21,8 +21,6 @@ __all__ = [
     "Coordinates",
     "Location",
     "SystemInfo",
-    "UserError",
-    "NoUserFound",
     "Case",
     "Platform",
     "KFCoords",
