+++ halpybot/packages/notify/__init__.py
@@ -13,9 +13,6 @@ from .notify import (
     list_sub_by_topic,
     subscribe,
     send_notification,
-    SNSError,
-    SubscriptionError,
-    NotificationFailure,
 )
 
 __all__ = [
@@ -23,7 +20,4 @@ __all__ = [
     "list_sub_by_topic",
     "subscribe",
     "send_notification",
-    "SNSError",
-    "SubscriptionError",
-    "NotificationFailure",
 ]
