+++ halpybot/packages/utils/__init__.py
@@ -13,8 +13,11 @@ from .utils import (
     language_codes,
     web_get,
     task_starter,
+    cache_prep,
+    sys_cleaner,
 )
 from .shorten import shorten
+from .decorators import CommandUtils
 
 __all__ = [
     "strip_non_ascii",
@@ -22,4 +25,7 @@ __all__ = [
     "shorten",
     "web_get",
     "task_starter",
+    "cache_prep",
+    "CommandUtils",
+    "sys_cleaner",
 ]
